Fix integer overflow in ElementCount() by using size_t with overflow checks
ElementCount() previously used a 32-bit signed integer (int)
to multiply tensor dimensions, which could overflow when
large dimensions were provided (e.g., [65536, 65536]).

This change:
- Changes the return type from int to size_t
- Adds overflow checks before multiplication
- Prevents undersized buffer allocations and memory corruption

BUG=544992926

+++ tensorflow/lite/micro/micro_utils.h
@@ -27,7 +27,7 @@ namespace tflite {
 
 // Returns number of elements in the shape array.
 
-int ElementCount(const TfLiteIntArray& dims);
+size_t ElementCount(const TfLiteIntArray& dims);
 
 size_t EvalTensorBytes(const TfLiteEvalTensor* tensor);
